requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
spacy==3.7.2
PyMuPDF==1.23.8
python-docx==1.1.0
python-multipart==0.0.6
